Add support for client side caching of GetTable and GetDatabase response
We have factored out a new interface(AWSGlueMetastore) between GlueMetastoreClientDelegate and AWSGlue, and the cache is implemented over this new interface.
For both GetTable and GetDatabase, we have separate instances of caches which can be enabled/tuned separately. Please refer to the 'Enabling client side caching for catalog' section in the README file for more details.